Share incoming body implementation between both client and server
