+++ crates/lune-std-net/src/shared/mod.rs
@@ -1,4 +1,5 @@
 pub mod headers;
 pub mod hyper;
+pub mod incoming;
 pub mod request;
 pub mod response;
